add auth

+++ api_node/controllers/cont_article.js
@@ -8,7 +8,7 @@ exports.ping = (req,res)=>{
 
 exports.get_article = (req,res)=>{
 
-    articleSchem.find()
+    articleSchem.find({userId : req.userId })
       .then(articles => res.status(200).json(articles) )
       .catch(err =>res.status(404).json({err: err.message}))
 }
@@ -16,8 +16,10 @@ exports.get_article = (req,res)=>{
 
 exports.post_article = (req,res)=>{
 
+    console.log('------------',req.userId)
     const article = new articleSchem({
         ...req.body,
+        userId : req.userId
     })
     article.save()
         .then(()=> res.status(201).json({ok : 'Le produit à ete ajouté'}))
@@ -26,14 +28,14 @@ exports.post_article = (req,res)=>{
 
 exports.update_article = (req,res)=>{
 
-    articleSchem.updateOne({_id: req.params.id},{...req.body})
+    articleSchem.updateOne({_id: req.params.id, userId : req.userId},{...req.body,userId : req.userId})
         .then(() => res.status(200).json({ok : 'Le produit a ete modifié '}) )
         .catch(err =>res.status(404).json({err : err.message}))
 }
 
 exports.delete_article = (req,res)=>{
 
-    articleSchem.deleteOne({_id: req.params.id})
+    articleSchem.deleteOne({_id: req.params.id, userId : req.userId})
         .then(() => res.status(200).json({ok : 'Le produit a ete supprimer'} )  )
         .catch(err =>res.status(404).json({err : err.message}))
 }
